Refactoring of some frontend related modules
* Move cache-control from Frontend.Handlers to Frontend
* Move the local formH/FormH/GetH/PostH to Frontend.Handlers.Combinators
* Add Frontend.Pages.Core for base layout, errors, basic validation
* Other minor improvements
