+++ thentos-tests/tests/Thentos/Frontend/StateSpec.hs
@@ -8,33 +8,28 @@
 
 module Thentos.Frontend.StateSpec where
 
-import Control.Lens ((^.), (%~))
-import Control.Monad (when)
-import Control.Monad.State (modify, gets, liftIO)
-import Data.Maybe (catMaybes, fromMaybe)
-import Data.String.Conversions (ST, LBS, SBS, cs)
 import Network.HTTP.Types (RequestHeaders, methodGet, methodPost)
 import Network.Wai (Application)
 import Network.Wai.Test (SResponse, simpleBody, simpleHeaders, simpleStatus)
-import Servant (Proxy(Proxy), ServerT, Capture, QueryParam, Post, Get, JSON, (:<|>)((:<|>)), (:>))
-import Test.Hspec (Spec, hspec, before, around, describe, it,
-                   shouldBe, shouldContain, shouldSatisfy, shouldNotSatisfy, pendingWith)
+import Servant (ServerT, Capture, QueryParam, Post, Get, JSON, (:<|>)((:<|>)), (:>))
+import Test.Hspec (Spec, hspec, before, around, describe, it, shouldBe, shouldContain,
+                   shouldSatisfy, shouldNotSatisfy, pendingWith)
 import Test.Hspec.Wai (request)
 import Servant.Server.Internal.ServantErr (errHTTPCode, errHeaders, errBody)
 
 import qualified Data.Attoparsec.ByteString.Char8 as AP
 import qualified Network.HTTP.Types.Status as HT
 import qualified Network.Wreq as Wreq
+import qualified Test.Hspec.Wai as HW
 
 import Thentos.Action.Types
-import Thentos.Config
-import Thentos.Frontend.Handlers.Combinators (redirect')
+import Thentos.Config hiding (getDefaultUser)
+import Thentos.Frontend.Handlers.Combinators
 import Thentos.Frontend.State
 import Thentos.Frontend.Types
-import Thentos.Types
-
+import Thentos.Prelude
 import Thentos.Test.Core
-
+import Thentos.Types
 
 tests :: IO ()
 tests = hspec spec
@@ -56,13 +51,13 @@ spec_frontendState = do
             errHTTPCode e `shouldBe` 404
             cs (errBody e) `shouldContain` ("<!DOCTYPE HTML>" :: String)
 
-    describe "the FAction monad, via hspec-wai" . before testApp $ do
+    describe "the FAction monad, on a test API, via hspec-wai" . before testApp $ do
         it "is initialized with empty message queue" $ do
-            resp <- request methodGet "" [] ""
+            resp <- HW.get ""
             liftIO $ simpleBody resp `shouldBe` "[]"
 
         it "posts accumulate into message queue" $ do
-            presp <- request methodPost "heya" [] ""
+            presp <- HW.post "heya" ""
             gresp <- request methodGet "" (getCookie presp) ""
             liftIO $ simpleBody gresp `shouldBe`
                 (cs . show . fmap show $ [FrontendMsgSuccess "heya"])
@@ -73,13 +68,13 @@ spec_frontendState = do
                 (cs . show . fmap show $ [FrontendMsgSuccess "ping", FrontendMsgSuccess "heya"])
 
         it "keeps state even when exceptions are thrown" $ do
-            presp <- request methodPost "heya?crash=True" [] ""
+            presp <- HW.post "heya?crash=True" ""
             gresp <- request methodGet "" (getCookie presp) ""
             liftIO $ simpleBody gresp `shouldBe`
                 (cs . show . fmap show $ [FrontendMsgSuccess "heya"])
 
         it "renders 404 correctly." $ do
-            resp <- request methodGet "/wef/yo" [] ""
+            resp <- HW.get "/wef/yo"
             liftIO $ HT.statusCode (simpleStatus resp) `shouldBe` 404
             liftIO $ pendingWith "rendering of implicit servant 404 (thrown by `HasServer (:>)`) cannot be customized."
             -- FIXME: it would be easy to get this to work with a catch-all combinator
@@ -151,9 +146,9 @@ testApi :: ServerT TestApi FAction
 testApi = post_ :<|> read_
   where
     post_ msg (fromMaybe False -> crash_) = do
-        modify $ fsdMessages %~ (FrontendMsgSuccess msg :)
+        fsdMessages %= (FrontendMsgSuccess msg :)
         when crash_ $ redirect' "/wef"
-    read_ = gets ((cs . show <$>) . (^. fsdMessages))
+    read_ = uses fsdMessages (cs . show <$>)
 
 testApp :: IO Application
 testApp = createActionState >>= serveFAction (Proxy :: Proxy TestApi) testApi
